Trying to go with a new direction with different min and max node implementations
I used tree class to handle different nodes
then a node will be the parent of maxnodes and the minnodes
what will they do is taking care of different types of nodes
but the problem stays with the repeatative nodes type. still no solution has been done in this matter.